backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyswisseph>=2.10.3
geopy>=2.4.1
timezonefinder>=6.4.0
pytz>=2024.1
pydantic>=2.6.0
pytest>=8.0.0
